Split blockchain invocation logic into new file
