+++ .env
@@ -3,5 +3,4 @@ ETHCONNECT_URL=http://127.0.0.1:5102
 ETHCONNECT_INSTANCE=/contracts/erc1155
 CONTRACT_ADDRESS=
 ETHCONNECT_TOPIC=token
-ETHCONNECT_PREFIX=fly
 AUTO_INIT=true
