+++ src/health/health.controller.spec.ts
@@ -16,7 +16,7 @@
 
 import { Test, TestingModule } from '@nestjs/testing';
 import { HealthCheckService, HttpHealthIndicator } from '@nestjs/terminus';
-import { TokensService } from '../tokens/tokens.service';
+import { BlockchainConnectorService } from '../tokens/blockchain.service';
 import { HealthController } from './health.controller';
 
 describe('HealthController', () => {
@@ -35,7 +35,7 @@ describe('HealthController', () => {
           useValue: jest.fn(),
         },
         {
-          provide: TokensService,
+          provide: BlockchainConnectorService,
           useValue: jest.fn(),
         },
       ],
