+++ src/tokens/tokens.service.spec.ts
@@ -18,6 +18,7 @@ import { HttpService } from '@nestjs/axios';
 import { Test, TestingModule } from '@nestjs/testing';
 import { EventStreamService } from '../event-stream/event-stream.service';
 import { EventStreamProxyGateway } from '../eventstream-proxy/eventstream-proxy.gateway';
+import { BlockchainConnectorService } from './blockchain.service';
 import { TokensService } from './tokens.service';
 
 describe('TokensService', () => {
@@ -38,6 +39,7 @@ describe('TokensService', () => {
     const module: TestingModule = await Test.createTestingModule({
       providers: [
         TokensService,
+        BlockchainConnectorService,
         {
           provide: HttpService,
           useValue: jest.fn(),
