+++ test/app.e2e-context.ts
@@ -11,10 +11,10 @@ import { EventStreamReply, EventBatch } from '../src/event-stream/event-stream.i
 import { EventStreamService } from '../src/event-stream/event-stream.service';
 import { EventStreamProxyGateway } from '../src/eventstream-proxy/eventstream-proxy.gateway';
 import { TokensService } from '../src/tokens/tokens.service';
+import { BlockchainConnectorService } from '../src/tokens/blockchain.service';
 
 export const BASE_URL = 'http://eth';
 export const INSTANCE_PATH = '/tokens';
-export const PREFIX = 'fly';
 export const TOPIC = 'tokentest';
 export const CONTRACT_ADDRESS = '0x888';
 
@@ -71,9 +71,8 @@ export class TestContext {
     await this.app.init();
 
     this.app.get(EventStreamProxyGateway).configure('url', TOPIC);
-    this.app
-      .get(TokensService)
-      .configure(BASE_URL, INSTANCE_PATH, TOPIC, PREFIX, '', '', CONTRACT_ADDRESS);
+    this.app.get(TokensService).configure(BASE_URL, INSTANCE_PATH, TOPIC, CONTRACT_ADDRESS);
+    this.app.get(BlockchainConnectorService).configure(BASE_URL, '', '');
 
     (this.app.getHttpServer() as Server).listen();
     this.server = request(this.app.getHttpServer());
